use favfactor for powerschedules also
