+++ libafl/src/corpus/testcase.rs
@@ -1,20 +1,14 @@
 //! The testcase is a struct embedded in each corpus.
 //! It will contain a respective input, and metadata.
 
-use alloc::string::{String, ToString};
+use alloc::string::String;
 use core::{convert::Into, default::Default, option::Option, time::Duration};
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    bolts::{serdeany::SerdeAnyMap, HasLen, HasRefCnt},
-    corpus::Corpus,
-    feedbacks::MapIndexesMetadata,
+    bolts::{serdeany::SerdeAnyMap, HasLen},
     inputs::Input,
-    schedulers::{
-        minimizer::{IsFavoredMetadata, TopRatedsMetadata},
-        powersched::{PowerSchedule, PowerScheduleMetadata},
-    },
-    state::{HasCorpus, HasMetadata},
+    state::HasMetadata,
     Error,
 };
 
@@ -57,11 +51,6 @@ where
         &mut self.metadata
     }
 }
-
-/// Constants for powerschedules
-const POWER_BETA: f64 = 1.0;
-const MAX_FACTOR: f64 = POWER_BETA * 32.0;
-const HAVOC_MAX_MULT: f64 = 64.0;
 
 /// Impl of a testcase
 impl<I> Testcase<I>
@@ -211,280 +200,6 @@ where
             executions,
             ..Testcase::default()
         }
-    }
-
-    /// Compute the `weight` used in weighted corpus entry selection algo
-    #[allow(clippy::cast_precision_loss, clippy::cast_lossless)]
-    pub fn compute_weight<S>(&self, state: &S) -> Result<f64, Error>
-    where
-        S: HasCorpus<I> + HasMetadata,
-    {
-        let mut weight = 1.0;
-        let psmeta = state
-            .metadata()
-            .get::<PowerScheduleMetadata>()
-            .ok_or_else(|| Error::KeyNotFound("PowerScheduleMetadata not found".to_string()))?;
-
-        let tcmeta = self
-            .metadata()
-            .get::<PowerScheduleTestcaseMetaData>()
-            .ok_or_else(|| Error::KeyNotFound("PowerScheduleTestData not found".to_string()))?;
-
-        // This means that this testcase has never gone through the calibration stage before1,
-        // In this case we'll just return the default weight
-        if tcmeta.fuzz_level() == 0 || psmeta.cycles() == 0 {
-            return Ok(weight);
-        }
-
-        let q_exec_us = self
-            .exec_time()
-            .ok_or_else(|| Error::KeyNotFound("exec_time not set".to_string()))?
-            .as_nanos() as f64;
-        let favored = self.has_metadata::<IsFavoredMetadata>();
-
-        let avg_exec_us = psmeta.exec_time().as_nanos() as f64 / psmeta.cycles() as f64;
-        let avg_bitmap_size = psmeta.bitmap_size() / psmeta.bitmap_entries();
-
-        let q_bitmap_size = tcmeta.bitmap_size() as f64;
-
-        match psmeta.strat() {
-            PowerSchedule::FAST | PowerSchedule::COE | PowerSchedule::LIN | PowerSchedule::QUAD => {
-                let hits = psmeta.n_fuzz()[tcmeta.n_fuzz_entry()];
-                if hits > 0 {
-                    weight *= libm::log10(f64::from(hits)) + 1.0;
-                }
-            }
-            // EXPLORE and EXPLOIT fall into this
-            _ => {}
-        }
-
-        weight *= avg_exec_us / q_exec_us;
-        weight *= libm::log2(q_bitmap_size) / (avg_bitmap_size as f64);
-
-        let tc_ref = match self.metadata().get::<MapIndexesMetadata>() {
-            Some(meta) => meta.refcnt() as f64,
-            None => 0.0,
-        };
-
-        let avg_top_size = state
-            .metadata()
-            .get::<TopRatedsMetadata>()
-            .ok_or_else(|| Error::KeyNotFound("TopRatedsMetadata not found".to_string()))?
-            .map()
-            .len() as f64;
-        weight *= 1.0 + (tc_ref / avg_top_size);
-
-        if favored {
-            weight *= 5.0;
-        }
-
-        // was it fuzzed before?
-        if tcmeta.fuzz_level() == 0 {
-            weight *= 2.0;
-        }
-
-        assert!(weight.is_normal());
-
-        Ok(weight)
-    }
-
-    /// Compute the `power` we assign to each corpus entry
-    #[inline]
-    #[allow(
-        clippy::cast_precision_loss,
-        clippy::too_many_lines,
-        clippy::cast_sign_loss
-    )]
-    pub fn calculate_score<S>(&self, state: &S) -> Result<usize, Error>
-    where
-        S: HasCorpus<I> + HasMetadata,
-    {
-        let psmeta = state
-            .metadata()
-            .get::<PowerScheduleMetadata>()
-            .ok_or_else(|| Error::KeyNotFound("PowerScheduleMetadata not found".to_string()))?;
-
-        let fuzz_mu = if psmeta.strat() == PowerSchedule::COE {
-            let corpus = state.corpus();
-            let mut n_paths = 0;
-            let mut v = 0.0;
-            for idx in 0..corpus.count() {
-                let n_fuzz_entry = corpus
-                    .get(idx)?
-                    .borrow()
-                    .metadata()
-                    .get::<PowerScheduleTestcaseMetaData>()
-                    .ok_or_else(|| {
-                        Error::KeyNotFound("PowerScheduleTestData not found".to_string())
-                    })?
-                    .n_fuzz_entry();
-                v += libm::log2(f64::from(psmeta.n_fuzz()[n_fuzz_entry]));
-                n_paths += 1;
-            }
-
-            if n_paths == 0 {
-                return Err(Error::Unknown(String::from("Queue state corrput")));
-            }
-
-            v /= f64::from(n_paths);
-            v
-        } else {
-            0.0
-        };
-
-        let mut perf_score = 100.0;
-        let q_exec_us = self
-            .exec_time()
-            .ok_or_else(|| Error::KeyNotFound("exec_time not set".to_string()))?
-            .as_nanos() as f64;
-
-        let avg_exec_us = psmeta.exec_time().as_nanos() as f64 / psmeta.cycles() as f64;
-        let avg_bitmap_size = psmeta.bitmap_size() / psmeta.bitmap_entries();
-
-        let favored = self.has_metadata::<IsFavoredMetadata>();
-        let tcmeta = self
-            .metadata()
-            .get::<PowerScheduleTestcaseMetaData>()
-            .ok_or_else(|| {
-                Error::KeyNotFound("PowerScheduleTestcaseMetaData not found".to_string())
-            })?;
-
-        if q_exec_us * 0.1 > avg_exec_us {
-            perf_score = 10.0;
-        } else if q_exec_us * 0.2 > avg_exec_us {
-            perf_score = 25.0;
-        } else if q_exec_us * 0.5 > avg_exec_us {
-            perf_score = 50.0;
-        } else if q_exec_us * 0.75 > avg_exec_us {
-            perf_score = 75.0;
-        } else if q_exec_us * 4.0 < avg_exec_us {
-            perf_score = 300.0;
-        } else if q_exec_us * 3.0 < avg_exec_us {
-            perf_score = 200.0;
-        } else if q_exec_us * 2.0 < avg_exec_us {
-            perf_score = 150.0;
-        }
-
-        let q_bitmap_size = tcmeta.bitmap_size() as f64;
-        if q_bitmap_size * 0.3 > avg_bitmap_size as f64 {
-            perf_score *= 3.0;
-        } else if q_bitmap_size * 0.5 > avg_bitmap_size as f64 {
-            perf_score *= 2.0;
-        } else if q_bitmap_size * 0.75 > avg_bitmap_size as f64 {
-            perf_score *= 1.5;
-        } else if q_bitmap_size * 3.0 < avg_bitmap_size as f64 {
-            perf_score *= 0.25;
-        } else if q_bitmap_size * 2.0 < avg_bitmap_size as f64 {
-            perf_score *= 0.5;
-        } else if q_bitmap_size * 1.5 < avg_bitmap_size as f64 {
-            perf_score *= 0.75;
-        }
-
-        if tcmeta.handicap() >= 4 {
-            perf_score *= 4.0;
-            // tcmeta.set_handicap(tcmeta.handicap() - 4);
-        } else if tcmeta.handicap() > 0 {
-            perf_score *= 2.0;
-            // tcmeta.set_handicap(tcmeta.handicap() - 1);
-        }
-
-        if tcmeta.depth() >= 4 && tcmeta.depth() < 8 {
-            perf_score *= 2.0;
-        } else if tcmeta.depth() >= 8 && tcmeta.depth() < 14 {
-            perf_score *= 3.0;
-        } else if tcmeta.depth() >= 14 && tcmeta.depth() < 25 {
-            perf_score *= 4.0;
-        } else if tcmeta.depth() >= 25 {
-            perf_score *= 5.0;
-        }
-
-        let mut factor: f64 = 1.0;
-
-        // COE and Fast schedule are fairly different from what are described in the original thesis,
-        // This implementation follows the changes made in this pull request https://github.com/AFLplusplus/AFLplusplus/pull/568
-        match psmeta.strat() {
-            PowerSchedule::EXPLORE => {
-                // Nothing happens in EXPLORE
-            }
-            PowerSchedule::EXPLOIT => {
-                factor = MAX_FACTOR;
-            }
-            PowerSchedule::COE => {
-                if libm::log2(f64::from(psmeta.n_fuzz()[tcmeta.n_fuzz_entry()])) > fuzz_mu
-                    && !favored
-                {
-                    // Never skip favorites.
-                    factor = 0.0;
-                }
-            }
-            PowerSchedule::FAST => {
-                if tcmeta.fuzz_level() != 0 {
-                    let lg = libm::log2(f64::from(psmeta.n_fuzz()[tcmeta.n_fuzz_entry()]));
-
-                    match lg {
-                        f if f < 2.0 => {
-                            factor = 4.0;
-                        }
-                        f if (2.0..4.0).contains(&f) => {
-                            factor = 3.0;
-                        }
-                        f if (4.0..5.0).contains(&f) => {
-                            factor = 2.0;
-                        }
-                        f if (6.0..7.0).contains(&f) => {
-                            if !favored {
-                                factor = 0.8;
-                            }
-                        }
-                        f if (7.0..8.0).contains(&f) => {
-                            if !favored {
-                                factor = 0.6;
-                            }
-                        }
-                        f if f >= 8.0 => {
-                            if !favored {
-                                factor = 0.4;
-                            }
-                        }
-                        _ => {
-                            factor = 1.0;
-                        }
-                    }
-
-                    if favored {
-                        factor *= 1.15;
-                    }
-                }
-            }
-            PowerSchedule::LIN => {
-                factor = (tcmeta.fuzz_level() as f64)
-                    / f64::from(psmeta.n_fuzz()[tcmeta.n_fuzz_entry()] + 1);
-            }
-            PowerSchedule::QUAD => {
-                factor = ((tcmeta.fuzz_level() * tcmeta.fuzz_level()) as f64)
-                    / f64::from(psmeta.n_fuzz()[tcmeta.n_fuzz_entry()] + 1);
-            }
-        }
-
-        if psmeta.strat() != PowerSchedule::EXPLORE {
-            if factor > MAX_FACTOR {
-                factor = MAX_FACTOR;
-            }
-
-            perf_score *= factor / POWER_BETA;
-        }
-
-        // Lower bound if the strat is not COE.
-        if psmeta.strat() == PowerSchedule::COE && perf_score < 1.0 {
-            perf_score = 1.0;
-        }
-
-        // Upper bound
-        if perf_score > HAVOC_MAX_MULT * 100.0 {
-            perf_score = HAVOC_MAX_MULT * 100.0;
-        }
-
-        Ok(perf_score as usize)
     }
 }
